refactor(jsx): split compile component summary
